libtracker-data: Added tracker:domainIndex property
In this commit the tracker:domainIndex property has no function yet. It'll
be implemented to represent a setting that will make the storage redundantly
store a property in the table of its domain, in order to allow placing an
index on the column for that table.

For example nmm:MusicPiece has a domain-index on nie:title. This will make
it store a redundent copy of nie:title in nmm:MusicPiece's SQL table,
while also having a copy in nie:InformationElement's table. The advantage is
a less complex query and the possibility to set an index on the redundant
column in nmm:MusicPiece's table.