SPARQL: Support graph retrieval using property functions
Property functions accept two extra optional arguments. The first
optional argument specifies the delimiter between multiple property
values. It is ignored for single-valued properties and it defaults to
comma.

If the second optional argument is present, the graph is returned in
addition to the property value. The graph is separated from the
property value by the delimiter specified in the argument.