libtracker-miner: Move writeback task removal to UPDATE monitor events
The last monitor event that's bound to happen from a writeback is a
UPDATED event caused by:

UPDATED(a) + MOVE(a->b) = MOVE(a->b) + UPDATED(b), caused by the delay
in handling the UPDATED so they're merged together. So don't remove
writeback tasks until we've gotten an update event for that file.